Update mlflow builder to also build the trainer image
Currently the mlflow builder is being used just as a means for providing
the Dockerfile to build the trainer image.

This change updates the mlflow builder to also build the trainer image,
besides, it also checks if an image with the same dependencies already
exists on the registry and in that case skips building the image allowing
previsouly built images to be reused.

Fixes: fuseml/fuseml#162
Partially fixes: fuseml/fuseml#59