kfserving: Update the wait condition to the new installer state
Now wait condition could be part of yaml file and no extra script
is needed.

Unfortunatelly `kubectl wait` cannot be used to wait for statefulsets
and `kubectl rollout` is not supported by fuseml-installer, so let's
replace it by the wait for all pods, that should be more or less
sufficient.