mlflow-builder: use miniconda image based on debian
The current base image used to build the trainer image is based on
alpine, which does not support glibc libraries. This commit changes
the base image to use a miniconda image that is based on Debian
instead. Also, removes the installation of git as it is already
included on the image.

Fixes: https://github.com/fuseml/fuseml/issues/262