OVMS predictor: model prediction type as step parameter
The prediction operation type featured by the TensorFlow API [1]
and reflected in the prediction URLs can take one of the `classify`,
`regress` or `predict` values.
It's too difficult to determine from the model which of these
operations are applicable to a given model, so we're modelling it as
a step parameter instead.

[1] https://www.tensorflow.org/tfx/serving/api_rest