Adapt the builder process for different kind of requirements
If requirements.txt file is found in the codeset directory, it is
converted to conda.yaml so that the same image can be used for training.

Additionally provide way to specify base image version. Some specific ML
dependencies are not available in the latest conda image so let user
have a chance to specify version of conda base image in the workflow file.