kfserving: accept RUNTIME_VERSION as workflow argument
There might be cases where the library version used in the training
cannot exactly match the image version used for inference. So give
user the chance to specify explicitly which version to use.

(For example when tensorflow is trained with 2.5.0 but there is no
tensorflow/serving image with such tag; there's 2.5.0-rc3 available)

Also, fix the awk command to work when tensorflow is listed more times
in requirements.txt (once for example as 'intel-tensorflow')