Include workflow name in the resources identification
When creating specific K8S and FuseML resources for inference services,
use the workflow name as part of the identification.

This way, when a codeset is run with different workflows, new resources
are created for each codeset-workflow combination instead of relying on
codeset information only, thus being recreated with the same name.