Predictor step improvements
1. list all input parameters explicitly as environment variables
in the Dockerfiles

2. add `app_name` parameter: implements an `app_name` input
parameter that can be set for all predictor steps that determines
the name of the FuseML application as well as all k8s resources
created for it. This allows users to create multiple applications
(i.e. prediction services) within a single workflow without resulting
in name clashing. It also allows users to control the same application
from multiple workflows. Setting the `app_name` input parameter
explicitly also removes the need to mount a codeset into the predictor
step.

3. add `verbose` parameter that controls verbosity

4. fix checking of existing folders in minio bucket