Model MLproject entrypoint arguments as workflow step parameters
The MLProject file also contains information about the entrypoints
(i.e. scripts to be executed), as well as the arguments that can
be passed to them. We haven't yet used that information: we assume
that a `main` entrypoint is always present and used by default and
we ignore the entrypoint arguments.

This change introduces a new feature that enables users to specify
values for `mlflow run` arguments in the form of workflow input
parameters.

The MLproject entrypoints and their arguments are converted into
environment variables for the trainer container images.
Users may now specify the following optional input parameters to the
MLFlow training workflow step:
- `mlflow_entrypoint` - an MLFlow project entrypoint to be executed
by the training workflow step (default: `main`).
- `mlflow_entrypoint_args` - list of explicit entrypoint arguments
to be supplied to the MLFlow project run, formatted as <name>=<value>
- `mlflow_experiment` - the name of the MLFlow experiment used when
running the MLFlow code in the training step (default: generated
from the codeset's name and project)
- `mlflow_artifact_path` - path relative to the the MLFlow project
run output artifact storage location to return as result (default:
`model`)

As an alternative to using `mlflow_entrypoint_args`, users may also
supply to the workflow trainer step input parameters that have the
same name as the entrypoint arguments. Their values will be passed
to the MLFlow project run.