MLFlow builder container image improvements
Define all workflow step arguments as Dockerfile env vars and
model new input parameters:

- `verbose`: displays verbose information during the step execution
when set to `true`
- `pull_registry`: can be used to set the container registry hostname
used by consumers when pulling the built trainer image. This can
be different than the `registry` input parameter that defines the
registry hostname used by the builder step when publishing the trainer
image to the registry. The defaults remain unchanged and are set to
match the requirements of using Trow as an internal container registry.